+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Added support for setting [PSR-18](https://www.php-fig.org/psr/psr-18/) HTTP client in
+  `ClientRaw`.
+
 
 ## [1.1.0](https://github.com/badoo/jira-client/compare/v1.0.0...v1.1.0) (2020-02-04)